# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Pitak"
        }
    ],
    "paths": {
...
        "/api/comment/create": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Write your comment",
                "operationId": "commentCreate",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "client_id": {
                                        "description": "write your client id(select option)",
                                        "type": "integer"
                                    },
                                    "driver_id": {
                                        "description": "write your driver id(select option)",
                                        "type": "integer"
                                    },
                                    "order_id": {
                                        "description": "write your order id(select option)",
                                        "type": "integer"
                                    },
                                    "text": {
                                        "description": "write your text message",
                                        "type": "string"
                                    },
                                    "score": {
                                        "description": "write your score",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/comment/get-comments?driver_id=21": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Finds Pets by status",
                "description": "Multiple status values can be provided with comma separated string",
                "operationId": "getComments",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available",
                            "enum": [
                                "available",
                                "pending",
                                "sold"
                            ]
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid status value"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Sanctum security laravel",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}